requirements.txt
streamlit
langchain
langchain-community
langchain-core
faiss-cpu
cohere
python-dotenv
fastapi
uvicorn
pydantic
together
langchain-together